Move DEBUG code paths from Main Delegate to Debugging Delegate
